+++ test/file/mold_assertions.cpp
@@ -81,9 +81,17 @@ SCENARIO("mold file::Interface ExistsException", "[PE-40]") {
  *
  */
 SCENARIO("mold file::Interface FolderExistsException", "[PE-40]") {
-  REQUIRE_THROWS_AS(file::ExistsException::assertion(".", __INFO__),
-                    file::ExistsException);
-  REQUIRE_THROWS_AS(file::ExistsException::assertion("..", __INFO__),
-                    file::ExistsException);
-  file::ExistsException::assertion("/>?/", __INFO__);
+  REQUIRE_THROWS_AS(
+      file::FolderExistsException::assertion(file::File(""), __INFO__),
+      file::FolderNotSpecifiedException);
+  REQUIRE_THROWS_AS(file::FolderExistsException::assertion(
+                        file::File("/usr/abc/*"), __INFO__),
+                    file::NotAFolderNameException);
+  REQUIRE_THROWS_AS(
+      file::FolderExistsException::assertion(file::File("/usr/abc"), __INFO__),
+      file::NotFoundException);
+  REQUIRE_THROWS_AS(file::FolderExistsException::assertion(
+                        file::File("/etc/hosts"), __INFO__),
+                    file::NotAFolderException);
+  file::FolderExistsException::assertion(file::File("/usr"), __INFO__);
 }