+++ changelog.md
@@ -5,8 +5,8 @@
 > and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).</br>
 > To add a tag to the git repository do:
 >
-> > git tag -a v8.4.0 -m "Dock/Molds Interfaces Framework"</br>
-> > git push origin v8.4.0
+> > git tag -a v8.5.0 -m "Added file::FolderExistsException exception"</br>
+> > git push origin v8.5.0
 >
 > > POPULAR LINKS:
 > > https://zed0.co.uk/clang-format-configurator/
@@ -25,6 +25,17 @@ IMPORTANT:
 
 ## [Unreleased]
 - replacement for filesystem include file 
+
+#### 8.5.0 (2022-10-16)
+### Added file::FolderExistsException exception 
+### Modified 
+- include/extras/language/interfaces.hpp
+### Changed
+- src/file/assertions.cpp (from src/file/exceptions.cpp)
+- test/fileinterface to test/file
+- changelog.md
+* we now have a convenient way of determining that a 
+* folder exists
 
 #### 8.4.0 (2022-09-20)
 ### Added 
@@ -791,7 +802,8 @@ IMPORTANT:
 ### Removed
 - nothing
 
-[unreleased]: https://github.com/perriera/extras/compare/v8.4.0..HEAD
+[unreleased]: https://github.com/perriera/extras/compare/v8.5.0..HEAD
+[8.5.0]: https://github.com/perriera/extras/compare/v8.4.0...v8.5.0
 [8.4.0]: https://github.com/perriera/extras/compare/v8.3.1...v8.4.0
 [8.3.1]: https://github.com/perriera/extras/compare/v8.3.0...v8.3.1
 [8.3.0]: https://github.com/perriera/extras/compare/v8.2.0...v8.3.0